+++ Project.toml
@@ -7,13 +7,20 @@ CoolProp_jll = "3351c21f-4feb-5f29-afb9-f4fcb0e27549"
 Markdown = "d6f4376e-aef5-505a-96c1-9c027394607a"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
+[weakdeps]
+Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
+
+[extensions]
+CoolPropUnitfulExt = "Unitful"
+
 [compat]
 CoolProp_jll = "6.6, 7.1"
 julia = "1.3"
-Unitful="1"
+Unitful = "1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
 [targets]
-test = ["Test"]
+test = ["Test", "Unitful"]
